Be more flexible in versions of required packages
"It is not considered best practice to use install_requires to pin dependencies to specific versions, or to specify sub-dependencies (i.e. dependencies of your dependencies). This is overly-restrictive, and prevents the user from gaining the benefit of dependency upgrades."
    --- https://packaging.python.org/en/latest/requirements/#install-requires-vs-requirements-files

Agate 1.3.1 is required because the `skip_lines` argument was added to `agate.Table.from_csv()` there. (1.3.0 would be required before that because the `printable` argument was added to `agate.Table.print_bars()`.)

As far as I can tell, nothing from Click 6.1+ is required for Bars to run, so the requirement has been reduced to >=6.0.